requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.18.0
datasets>=2.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0
wordcloud>=1.8.0
tensorboard>=2.8.0
opencv-python>=4.5.0
requests>=2.25.0
jupyter>=1.0.0
ipython>=7.0.0 